+++ packages/qti-convert-core/src/qti-upgrader/index.ts
@@ -1,1 +1,2 @@
 export { upgradeQti2toQti3 } from './convert-qti2-to-qti3';
+export * from './extract-shared-stimuli';